Combined two calls to blocking I2C function. Disabled retrieval of magnetometer data while not in use.

Performance Stats:
| STATE | DELTA | SAMPLING RATE |
| -------- | -------- | -------- |
| Flight | 4ms -> 3ms | 250hz -> 333hz |
| Launch Detect | 4ms -> 3ms | 250hz -> 333hz |

Closes #47.